Fix fsm and add order with mouse

+++ Simple IA/Assets/Scripts/Resource/ResourceCharge.cs
@@ -1,15 +1,13 @@
-using UnityEngine;
-
 [System.Serializable]
 public class ResourceCharge
 {
     public ResourceCharge(){}
-    public ResourceCharge (ResourceType resourceType, int resourceAmount)
+    public ResourceCharge (ResourceType resourceType, float resourceAmount)
     {
         this.resourceType = resourceType;
         this.resourceAmount = resourceAmount;
     }
 
     public ResourceType resourceType;
-    public int resourceAmount;
+    public float resourceAmount;
 }
